mini-apps/requirements-simple.txt
# Dependências básicas para ambos os apps
requests
python-socketio
# Versões compatíveis para evitar erros
Pillow==9.5.0
pyscreeze==0.1.29
pyautogui==0.9.54



